Protect against missing user parameters
Since the source code relies on many user-supplied parameters and nothing prevents the user from neglecting these, we have added a check for the existence of each of these user parameters before grabbing their values and dealt with what happens if they don't exist. When reasonable, a default value is set and a warning issued. In limited cases, this will cause an error to be thrown.