Integrated with rules.
No longer need to specify moa and moa parameters in custom_data. Instead, put all this in rules and PhysiPKPD will pull this info to identify combinations of (cell_definition, substrate) that form PD models and update damage for those. PhysiPKPD will turn on tracking of internalized agents if the user sets it to false. Custom.cpp code is no longer necessary to have the phenotypes reset to base rate.