Updated sample projects to work with the rules integration.
All user_parameters and custom_data (except damage variables) are now in their studio-defined places. Studio-defined means where I made changes to studio that put them in certain places in the xml.

Confluence example is temporarily deprecated as I don't care enough about that example to work on it.

SBML in a PK needs to be fully integrated with studio so the user can supply the name of the SBML filename (and it must be located in config).