More protections for editing files.
Now check if the lines to be added are already present before adding them. In Makefile and main.cpp, just check if the first line to be added is present and then assume the rest are the same. Warnings are issued if it is found and the rest are being skipped.